Consolidate blueprint-integration back into a single project, split HostApplicationContext
