--- blueprint-integration/blueprint-integration-51/build.gradle
@@ -1,3 +1,0 @@
-ext {
-    excludedSharedFiles = ["HostApplicationContext.java", "AutowireBeanFactory.java"]
-}
