--- blueprint-integration/blueprint-integration-52/build.gradle
@@ -1,3 +1,0 @@
-ext {
-    excludedSharedFiles = ["HostApplicationContext.java", "AutowireBeanFactory.java"]
-}
