+++ blueprint-integration/src/main/java/com/github/dynamicextensionsalfresco/blueprint/OsgiAutowireBeanFactory.java
