+++ blueprint-integration/src/main/resources/META-INF/spring.schemas
