+++ blueprint-integration/src/main/resources/META-INF/spring/extender/extender-context.xml
