+++ blueprint-integration/src/main/resources/org/springframework/beans/factory/xml/spring-beans-2.5.xsd
