+++ blueprint-integration/src/main/resources/org/springframework/beans/factory/xml/spring-beans-3.0.xsd
