+++ blueprint-integration/src/main/resources/org/springframework/beans/factory/xml/spring-tool-2.0.xsd
