+++ blueprint-integration/src/main/resources/org/springframework/beans/factory/xml/spring-tool-3.0.xsd
