+++ blueprint-integration/src/main/resources/org/springframework/beans/factory/xml/spring-util-2.0.xsd
