+++ blueprint-integration/src/main/resources/org/springframework/beans/factory/xml/spring-util-3.0.xsd
