+++ blueprint-integration/src/test/java/com/github/dynamicextensionsalfresco/blueprint/DynamicExtensionsApplicationContextTest.java
