+++ blueprint-integration/src/test/java/com/github/dynamicextensionsalfresco/osgi/spring/Example.java
