+++ blueprint-integration/src/test/resources/com/github/dynamicextensionsalfresco/osgi/spring/AutowireBeanFactoryTest-child-context.xml
