+++ blueprint-integration/src/test/resources/com/github/dynamicextensionsalfresco/osgi/spring/AutowireBeanFactoryTest-parent-context.xml
