+++ blueprint-integration/src/test/resources/com/github/dynamicextensionsalfresco/osgi/spring/BundleServiceProxyTest.xml
