+++ settings.gradle
@@ -13,13 +13,16 @@ include 'annotations-runtime:scheduler-quartz-2'
 
 include 'alfresco-integration'
 
+include 'blueprint-integration'
+include 'blueprint-integration:spring-3'
+include 'blueprint-integration:spring-5'
+
 include 'gradle-plugin'
 
 
 def alfrescoDependentModules =
         [
                 'alfresco-dynamic-extensions-repo',
-                'blueprint-integration',
                 'blueprint-override',
                 'control-panel'
         ]
